+++ pkg/vm_snapshot_analysis/bin/utils.dart
@@ -2,10 +2,13 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-import '_file_stub.dart'
-    if (dart.library.html) '_file_web.dart'
-    if (dart.library.io) '_file_vm.dart';
+import 'dart:convert';
+import 'dart:io';
 
-Future<Object> loadJson(dynamic file) {
-  return loadJsonFromFile(file);
+Future<Object> loadJsonFromFile(File input) async {
+  return await input
+      .openRead()
+      .transform(utf8.decoder)
+      .transform(json.decoder)
+      .first;
 }