+++ pkg/vm_snapshot_analysis/CHANGELOG.md
@@ -1,7 +1,11 @@
 # Changelog
 
 ## 0.5.0
-
+- Remove `dart:io` dependency from package `lib`, and move commands to `bin`.
+- Replace `loadProgramInfo` util method with `loadProgramInfoFromJson`, which
+expects an `Object` parameter instead of a `File` parameter.
+- `buildComparisonTreemap` now expects two `Object` parameters for `oldJson` and
+`newJson` instead of two `File` parameters.
 - `compare` command now prints difference breakdown by node type when this
 information is available.